requirements.txt
tqdm
python-dotenv
langchain
langchain-openai
openai
